requirements.txt
spacy>=3.5.0
spacy-llm>=0.6.0
python-dotenv>=1.0.0
transformers>=4.30.0
torch>=2.0.0
accelerate>=0.20.0
safetensors>=0.3.0
watchdog>=3.0.0
sentencepiece>=0.1.99
protobuf>=3.20.0
google-cloud-translate>=3.11.0
mcp>=0.9.0
pyperclip>=1.8.2
google-adk>=1.17.0
google-genai>=0.2.0
